backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
pyjwt==2.8.0
python-dotenv==1.0.0
werkzeug==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==12.1.0